[refactor] errorMessage 및 exception 변경 (#7)
