+++ src/main/java/org/recordy/server/auth/service/impl/apple/AppleClaimsValidator.java
@@ -2,7 +2,7 @@ package org.recordy.server.auth.service.impl.apple;
 
 import io.jsonwebtoken.Claims;
 import org.recordy.server.auth.exception.AuthException;
-import org.recordy.server.auth.message.ErrorMessage;
+import org.recordy.server.common.message.ErrorMessage;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
