+++ src/main/java/org/recordy/server/auth/service/impl/apple/AppleIdentityTokenParser.java
@@ -13,7 +13,7 @@ import java.security.PublicKey;
 import java.util.Base64;
 import java.util.Map;
 import org.recordy.server.auth.exception.AuthException;
-import org.recordy.server.auth.message.ErrorMessage;
+import org.recordy.server.common.message.ErrorMessage;
 import org.springframework.stereotype.Component;
 
 @Component
