+++ src/main/java/org/recordy/server/common/exception/RecordyException.java
@@ -1,13 +1,15 @@
 package org.recordy.server.common.exception;
 
+import lombok.Getter;
 import org.recordy.server.common.message.ErrorMessage;
 
-public class CommonException extends RuntimeException{
+@Getter
+public class RecordyException extends RuntimeException {
 
     private final ErrorMessage errorMessage;
-    public CommonException(ErrorMessage errorMessage){
+
+    public RecordyException(ErrorMessage errorMessage) {
         super(errorMessage.getMessage());
         this.errorMessage = errorMessage;
     }
-
 }
